backend/pyproject.toml
[project]
name = "checklist-backend"
version = "0.1.0"
description = "CheckList backend API with FastAPI on Modal"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "modal>=0.57.0",
    "weaviate-client>=4.4.0",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    "sentence-transformers>=2.3.0",
    "pymupdf>=1.23.0",
    "transformers==4.51.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
]